@@ -1,9 +1,9 @@
 module github.com/Akagi201/utilgo
 
-go 1.17
+go 1.18
 
 require (
-	github.com/BurntSushi/toml v0.4.1
+	github.com/BurntSushi/toml v1.0.0
 	github.com/fatih/structs v1.1.0
 	github.com/grooveshark/golib v0.1.0
 	github.com/jessevdk/go-flags v1.5.0
@@ -13,10 +13,13 @@ require (
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/net v0.0.0-20211101193420-4a448f8816b3
 	gopkg.in/yaml.v2 v2.4.0
+	gorm.io/gorm v1.22.5
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.4 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/sys v0.0.0-20210423082822-04245dca01da // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
