@@ -5,33 +5,33 @@ import "sync"
 
 // Set safe map
 type Set struct {
-	m map[interface{}]struct{}
+	m map[any]struct{}
 	sync.RWMutex
 }
 
 // New new set
 func New() *Set {
 	return &Set{
-		m: make(map[interface{}]struct{}),
+		m: make(map[any]struct{}),
 	}
 }
 
 // Add add
-func (s *Set) Add(item interface{}) {
+func (s *Set) Add(item any) {
 	s.Lock()
 	defer s.Unlock()
 	s.m[item] = struct{}{}
 }
 
 // Remove deletes the specified item from the map
-func (s *Set) Remove(item interface{}) {
+func (s *Set) Remove(item any) {
 	s.Lock()
 	defer s.Unlock()
 	delete(s.m, item)
 }
 
 // Has looks for the existence of an item
-func (s *Set) Has(item interface{}) bool {
+func (s *Set) Has(item any) bool {
 	s.RLock()
 	defer s.RUnlock()
 	_, ok := s.m[item]
@@ -47,7 +47,7 @@ func (s *Set) Len() int {
 func (s *Set) Clear() {
 	s.Lock()
 	defer s.Unlock()
-	s.m = make(map[interface{}]struct{})
+	s.m = make(map[any]struct{})
 }
 
 // IsEmpty checks for emptiness
@@ -56,17 +56,17 @@ func (s *Set) IsEmpty() bool {
 }
 
 // Each call fn on each element in the Set
-func (s *Set) Each(fn func(interface{})) {
+func (s *Set) Each(fn func(any)) {
 	for e := range s.m {
 		fn(e)
 	}
 }
 
 // List Set returns a slice of all items
-func (s *Set) List() []interface{} {
+func (s *Set) List() []any {
 	s.RLock()
 	defer s.RUnlock()
-	var list []interface{}
+	var list []any
 	for item := range s.m {
 		list = append(list, item)
 	}
